// frontend/types/notificationTypes.ts
export enum NotificationEvents {
    AuctionOutbidded = 'auction_outbidded',
    AuctionReceivedBid = 'auction_recieved_bid',
    AuctionBidWon = 'auction_bid_won',
    AuctionBidLost = 'auction_bid_lost',
    AuctionEndingSoon = 'auction_ending_soon',
    AuctionOwningEnded = 'auction_owning_ended',
    AuctionWatchingReceivedBid = 'auction_watching_recieved_bid',
    AuctionWatchingEnded = 'auction_watching_ended',
    AuctionWatchingEndingSoon = 'auction_watching_ending_soon',
}